Fix bug #1356: .obj loading bug in picolib
There's a bug in picolib when parsing the first group in an .obj file.

The library tries to set the surface name to a surface pointer
that is still NULL.

This gets silently ignored but the problem reappears later
in a warning message when the material or a face is parsed.